Rmove leagacy cursor component
