+++ components/theCursor/theCursor.module.css
@@ -4,13 +4,21 @@
 
 .TheCursor {
   position: fixed;
-  z-index: 9999;
+  z-index: 100;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  pointer-events: none;
+}
+
+.TheCursor__pointer {
+  position: absolute;
+  top: 0;
+  left: 0;
   width: 40px;
   height: 40px;
-  pointer-events: none;
-  transform: translate3d(-50%, -50%, 0);
-  transition: all 150ms ease;
-  transition-property: opacity, transform;
+  transform: translate(-50%, -50%) translate3d(0, 0, 0);
 
   &::before,
   &::after {
@@ -21,7 +29,7 @@
     height: 40px;
     background-color: $red;
     content: '';
-    transition: transform 150ms ease;
+    transition: transform 150ms, opacity 150ms ease;
   }
 
   &::before {
@@ -32,17 +40,11 @@
     transform: rotate(-45deg);
   }
 
-  &.TheCursor--hidden {
+  .TheCursor--hidden & {
     opacity: 0;
   }
 
-  &.TheCursor--clicked {
-    transform: translate3d(-50%, -50%, 0) scale(0.8);
-  }
-
-  &.TheCursor--hovered {
-    transform: translate3d(-50%, -50%, 0) scale(1.2);
-
+  .TheCursor--hovered & {
     &::before {
       transform: rotate(90deg);
     }
