--- components/theCursorV2/theCursorV2.module.css
@@ -1,56 +1,0 @@
-/* ------------------------------------ *\
-  #TheCursor
-\* ------------------------------------ */
-
-.TheCursor {
-  position: fixed;
-  z-index: 100;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  pointer-events: none;
-}
-
-.TheCursor__pointer {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 40px;
-  height: 40px;
-  transform: translate(-50%, -50%) translate3d(0, 0, 0);
-
-  &::before,
-  &::after {
-    position: absolute;
-    top: 0;
-    left: 50%;
-    width: 2px;
-    height: 40px;
-    background-color: $red;
-    content: '';
-    transition: transform 150ms, opacity 150ms ease;
-  }
-
-  &::before {
-    transform: rotate(45deg);
-  }
-
-  &::after {
-    transform: rotate(-45deg);
-  }
-
-  .TheCursor--hidden & {
-    opacity: 0;
-  }
-
-  .TheCursor--hovered & {
-    &::before {
-      transform: rotate(90deg);
-    }
-
-    &::after {
-      transform: rotate(0);
-    }
-  }
-}
